fixup! Replace bwlist by bw_file and terminator condition
Windows compilation fails because time_t is not a long int.

Instead, cast time_t back to an unsigned long, because that's
what we parsed it as.

+++ src/or/dirserv.c
@@ -2660,7 +2660,8 @@ dirserv_read_measured_bandwidths(const char *from_file,
   /* If timestamp was correct and bw_file_headers is not NULL,
    * add timestamp to bw_file_headers */
   if (bw_file_headers)
-    smartlist_add_asprintf(bw_file_headers, "timestamp=%ld", file_time);
+    smartlist_add_asprintf(bw_file_headers, "timestamp=%lu",
+                           (unsigned long)file_time);
 
   if (routerstatuses)
     smartlist_sort(routerstatuses, compare_vote_routerstatus_entries);
